# feat: Use binary search in `GetPrecompileModuleByAddress`