Benchmarks w/ 1000 registered modules (benchmarks included in pr):

```
goos: darwin
goarch: arm64
pkg: github.com/ethereum/go-ethereum/precompile/modules
                                │   old.txt    │              new.txt               │
                                │    sec/op    │   sec/op     vs base               │
GetPrecompileModuleByAddress-12   1460.5n ± 1%   319.1n ± 1%  -78.15% (p=0.002 n=6)

                                │  old.txt   │            new.txt            │
                                │    B/op    │    B/op     vs base           │
GetPrecompileModuleByAddress-12   7.000 ± 0%   7.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal

                                │  old.txt   │            new.txt            │
                                │ allocs/op  │ allocs/op   vs base           │
GetPrecompileModuleByAddress-12   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal
```

resolves #29